// src/interfaces/SenateResults.ts
export default interface SenateResults {
    id: number;
    name: string;
    numberOfVotes: number;
    votesForKO: number;
    votesForLEW: number;
    votesForTD: number;
    votesForKONF: number;
    votesForPIS: number;
    votesForBS: number;
    votesForMN: number;
    votesForSenatePact: number;
    votesForRightWingPact: number;
}